clean up some dead romdisk code
use mockingboard VIA timer 2 to repair keyboard state machine if parsing fails for any reason
reset a timer for 400us for every ps/2 keyboard clock pulse.
when timer elapses, call irq and reset the kb state machine.

test with U4 music as it uses mockingboard which uses the interrupts
U4 overwrites the irq routine, so it won't do anything, but setting the timer in the nmi will cause the irq interrupt to fire